+++ docs/source/conf.py
@@ -46,9 +46,9 @@ copyright = '2019, S. Domanskyi, C. Piermarocchi, G. Mias'
 author = 'S. Domanskyi, C. Piermarocchi, G. Mias'
 
 # The short X.Y version
-version = '1.1.2'
+version = '1.2.3'
 # The full version, including alpha/beta/rc tags
-release = '1.1.2'
+release = '1.2.3'
 
 
 # -- General configuration ---------------------------------------------------